Initial Work Towards 64-bit patch Still likely will have issues if you try to build 64 Filesystem and model code are likely big culprits too

+++ codemp/qcommon/cm_polylib.cpp
@@ -248,11 +248,11 @@ CopyWinding
 */
 winding_t	*CopyWinding (winding_t *w)
 {
-	int			size;
+	intptr_t	size;
 	winding_t	*c;
 
 	c = AllocWinding (w->numpoints);
-	size = (int)((winding_t *)0)->p[w->numpoints];
+	size = (intptr_t) ((winding_t *)0)->p[w->numpoints];
 	Com_Memcpy (c, w, size);
 	return c;
 }
